+++ src/main/java/dopamine/backend/jwt/controller/JwtController.java
@@ -1,17 +1,13 @@
 package dopamine.backend.jwt.controller;
 
-import com.google.gson.JsonObject;
 import dopamine.backend.jwt.dto.KakaoUserInfo;
 import dopamine.backend.jwt.response.JwtResponse;
 import dopamine.backend.jwt.service.JwtService;
-import dopamine.backend.jwt.service.OauthService;
 import dopamine.backend.member.entity.Member;
 import dopamine.backend.member.mapper.MemberMapper;
-import dopamine.backend.member.response.MemberResponseDto;
 import dopamine.backend.member.service.MemberService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -24,7 +20,6 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 public class JwtController {
 
-    private final OauthService oauthService;
     private final MemberService memberService;
     private final JwtService jwtService;
     private final MemberMapper memberMapper;
@@ -33,8 +28,8 @@ public class JwtController {
     public ResponseEntity kakaoCallback(@RequestParam String code) {
 
         // 유저 정보 얻기
-        String kakaoAccessToken = oauthService.getKakaoAccessToken(code);
-        KakaoUserInfo kakaoUserInfo = oauthService.getKakaoUserInfo(kakaoAccessToken);
+        String kakaoAccessToken = jwtService.getKakaoAccessToken(code);
+        KakaoUserInfo kakaoUserInfo = jwtService.getKakaoUserInfo(kakaoAccessToken);
 
         // 해당 kakao ID를 가진 Member 반환
         Member member = memberService.findMemberByKakaoId(kakaoUserInfo.getKakaoId());
@@ -48,12 +43,5 @@ public class JwtController {
                 .accessToken(accessToken)
                 .refreshToken(refreshToken).build();
         return new ResponseEntity<>(jwtResponse, HttpStatus.CREATED);
-    }
-
-    @GetMapping("/member")
-    public ResponseEntity returnMemberDetail(@RequestHeader("Authorization") String accessToken) {
-        Member member = jwtService.getMemberFromAccessToken(accessToken);
-        MemberResponseDto response = memberMapper.memberToMemberResponseDto(member);
-        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
